Created basic Network base class
